+++ src/checkers/php.rs
@@ -5,7 +5,7 @@
 use std::collections::HashMap;
 
 use super::HttpChecker;
-use crate::models::{Finding, Technology, UrlResponse};
+use crate::models::{Finding, Technology, UrlRequestType, UrlResponse};
 use log::{info, trace};
 use regex::Regex;
 
@@ -100,6 +100,13 @@ impl<'a> HttpChecker for PHPChecker<'a> {
     fn check_http(&self, data: &[UrlResponse]) -> Option<Finding> {
         trace!("Running PHPChecker::check_http()");
         for url_response in data {
+            // JavaScript files could be hosted on a different server
+            // Don't check the JavaScript files to avoid false positive,
+            // Check only the "main" requests.
+            if url_response.request_type != UrlRequestType::Default {
+                continue;
+            }
+
             let header_finding = self.check_http_headers(url_response);
             if header_finding.is_some() {
                 return header_finding;